Update shelfice_solve4fluxes to match variable names
The variable for the velocity of meltwater flux through the boundary layer was called *w_I* and is now called *w_B*.  The reason for the change is to match the document 'the three-equation model v8' that describes the model used to calculate melting.  the document will be included in the next ecco release.